Force memberaudit updates when a character applies
When a member applies, a memberaudit update is immediately queued for
that character. Unless a staff member looks at the application within
seconds of that action, this more or less solves the stale skill issue.

Because we're running MA 2.10, this is still a bit heavy-handed, since
we run a *full* character update every time, instead of just pulling
skills and corp/alliance. That can be easily fixed if MA is updated.