Add corp/alliance checks on application.
Added corp/alliance checks to the application action.

Refactored the ALLOWED_ALLIANCES out to make room for more complex
checks (e.g., NPC corps for Combat Comms, which will come later).

Cleaned up and compacted the user-side applications interface. I
added corp and alliance info (to make clear what Auth thinks the
character is when it reports an invalid alliance). While I was in
the area, I cleaned up some of the formatting so that it doesn't
sprawl so much.