Disabling the sync-groups-with-ACL button.
I thought better of this. During the grace period, we don't want this
button active. We want people added to the ACL to be added to the
appropriate groups, but we *don't* want people removed from those
groups yet.

Once the grace period is over, this button can be uncommented.
Pushing it will cleanse the ACL-associated groups from anyone not on
the ACL.