Fix ACL view for characters without mains.
Adds a new sub-group of members internally called "orphans" which
represents characters without a valid main (not characters with an
out-of-corp main). When AA can't retrieve a main, it returns a
sentinel value which it assumes is a deleted character. We check this
now and correctly flag these as needing attention.

Fixes #63.