Rework ACL UI.
Changed from a nested view to a flat list, largely because there were
problems when parents didn't exist. We do sort the list mostly in the
same fashion, so it should look very similar. Also indirectly provides
a search function, since you can just cntl-F it. (I'm not going to mark
the corresponding feature request as fixed, though. There should be an
actual filter for that.)
Fixes #31.

Apparently the reject function never worked. It looked for an attribute
that didn't exist.
Fixes #66.

Formatting changes to ACL text dump to remove some extraneous lines.