Move to only using eve_character.character_id
Previously, the code used both eve_character.id (an automatically-created,
internal django integer) and eve_character.character_id (the EVE integer
assigned by CCP). Thanks to django's questionable decision to allow the
appending of `_id` as a shortcut to referencing `.id`, this led to some truly
amusing templates where "main_character_id" was a user-supplied variable
that referenced an eve_character.character_id but "eve_character_id" in a
similar context was an eve_character.id. Certain view endpoints did the same.
This was the source of the bug in the reject view, as it was receiving id's
from some links and character_id's in others.

Well, no longer. Everything has been moved to using character_id's. There's a
minor performance tradeoff to that decision, but it won't affect anything
unless we reach tens or hundreds of thousands of members. That'd be nice.